django_material_demo/requirements.txt
asgiref==3.5.2
boto3==1.24.80
botocore==1.27.80
Django==4.1
django-filter==22.1
django-material==1.10.0
django-s3direct==2.0.3
django-storages==1.13.1
jmespath==1.0.1
Pillow==9.2.0
psycopg2==2.9.3
python-dateutil==2.8.2
python-dotenv==0.20.0
PyYAML==6.0
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
